Move FSFile to separate module and add some reader type annotations
